[ACCoRD] Update DAIDALUS version

+++ Modules/ACCoRD/src/ErrorLog.cpp
@@ -3,7 +3,7 @@
  *
  * Contact: Jeff Maddalon (j.m.maddalon@nasa.gov)
  *
- * Copyright (c) 2011-2018 United States Government as represented by
+ * Copyright (c) 2011-2019 United States Government as represented by
  * the National Aeronautics and Space Administration.  No copyright
  * is claimed in the United States under Title 17, U.S.Code. All Other
  * Rights Reserved.
@@ -62,7 +62,7 @@ void ErrorLog::addError(const string& msg) {
 	}
 	size++;
 	if (size > limit) {
-		int idx = message.find("\n");
+		size_t idx = message.find("\n");
 		message.erase(0, idx);
 		message.insert(0,"[...] ");
 		size--;
@@ -86,7 +86,7 @@ void ErrorLog::addWarning(const string& msg) {
 	}
 	size++;
 	if (size > limit) {
-		int idx = message.find("\n");
+		size_t idx = message.find("\n");
 		message.erase(0, idx);
 		message.insert(0,"[...] ");
 		size--;
